Reverses changes on internal browser
