+++ src/beagle-view/navigator/internal-navigator.ts
@@ -17,7 +17,9 @@
 import cloneDeep from 'lodash/cloneDeep'
 import last from 'lodash/last'
 import nth from 'lodash/nth'
+import find from 'lodash/find'
 import BeagleNavigationError from 'error/BeagleNavigationError'
+import logger from 'logger'
 import findLastIndex from 'lodash/findLastIndex'
 import {
   BeagleNavigator,
@@ -27,7 +29,7 @@ import {
   NavigationListener,
   NavigationController,
 } from './types'
-import { isRouteIdentifiedBy, runListeners } from './navigator.commons'
+import { isRouteIdentifiedBy } from './navigator.commons'
 
 const createBeagleNavigator = (
   navigationControllers?: Record<string, NavigationController>,
@@ -36,7 +38,17 @@ const createBeagleNavigator = (
   let navigation: Stack[] = initialValue ? cloneDeep(initialValue) : []
   let isNavigationInProgress = false
   let isDestroyed = false
+  const defaultNavigationController = find(navigationControllers, { default: true }) || {}
   const listeners: NavigationListener[] = []
+
+  function getNavigationController(controllerId?: string) {
+    if (!controllerId) return defaultNavigationController
+    if (!navigationControllers || !navigationControllers[controllerId]) {
+      logger.warn(`No navigation controller with id ${controllerId} has been found. Using the default navigation controller.`)
+      return defaultNavigationController
+    }
+    return navigationControllers[controllerId]
+  }
 
   function subscribe(listener: NavigationListener) {
     listeners.push(listener)
@@ -77,7 +89,13 @@ const createBeagleNavigator = (
     return route
   }
 
-  async function navigate(
+  function runListeners(route: Route) {
+    const controllerId = navigation.length ? getCurrentStack().controllerId : undefined
+    const navigationController = getNavigationController(controllerId)
+    return Promise.all(listeners.map(l => l(route, navigationController)))
+  }
+
+    async function navigate(
     type: NavigationType,
     route?: Route | string,
     controllerId?: string,
@@ -88,7 +106,7 @@ const createBeagleNavigator = (
           throw new BeagleNavigationError(`Invalid route for pushStack. Expected: Route object. Received: ${route}.`)
         }
 
-        await runListeners(route, listeners, controllerId, navigationControllers)
+        await runListeners(route)
         navigation.push({ routes: [route], controllerId })
       },
 
@@ -98,7 +116,7 @@ const createBeagleNavigator = (
         }
 
         const route = last(getPreviousStack().routes)!
-        await runListeners(route, listeners, controllerId, navigationControllers)
+        await runListeners(route)
         navigation.pop()
       },
 
@@ -107,7 +125,7 @@ const createBeagleNavigator = (
           throw new BeagleNavigationError(`Invalid route for pushView. Expected: Route object. Received: ${route}.`)
         }
 
-        await runListeners(route, listeners, controllerId, navigationControllers)
+        await runListeners(route)
         if (navigation.length === 0) navigation.push({ routes: [] })
         getCurrentStack().routes.push(route)
       },
@@ -117,7 +135,7 @@ const createBeagleNavigator = (
           throw new BeagleNavigationError('It was not possible to pop a view because Beagle Navigator has not more than one recorded route')
         }
 
-        await runListeners(getPreviousRoute(), listeners, controllerId, navigationControllers)
+        await runListeners(getPreviousRoute())
         const currentStack = getCurrentStack()
         currentStack.routes.pop()
         if (currentStack.routes.length <= 0) navigation.pop()
@@ -131,7 +149,7 @@ const createBeagleNavigator = (
         const currentStack = getCurrentStack()
         const index = findLastIndex(currentStack.routes, r => isRouteIdentifiedBy(r, route))
         if (index === -1) throw new BeagleNavigationError('The route does not exist in the current stack')
-        await runListeners(currentStack.routes[index], listeners, controllerId, navigationControllers)
+        await runListeners(currentStack.routes[index])
         currentStack.routes.splice(index + 1)
       },
 
@@ -140,7 +158,7 @@ const createBeagleNavigator = (
           throw new BeagleNavigationError(`Invalid route for pushView. Expected: Route object. Received: ${route}.`)
         }
 
-        await runListeners(route, listeners, controllerId, navigationControllers)
+        await runListeners(route)
         navigation.pop()
         navigation.push({ routes: [route], controllerId })
       },
@@ -150,7 +168,7 @@ const createBeagleNavigator = (
           throw new BeagleNavigationError(`Invalid route for resetApplication. Expected: Route object. Received: ${route}.`)
         }
 
-        await runListeners(route, listeners, controllerId, navigationControllers)
+        await runListeners(route)
         navigation = [{ routes: [route], controllerId }]
       },
     }
